[Update] 카테고리 수정 및 삭제 로직 수정 - 일부 코드 리팩토링

+++ src/main/java/keapoint/onlog/post/service/CategoryService.java
@@ -13,8 +13,8 @@ import keapoint.onlog.post.repository.TopicRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import keapoint.onlog.post.dto.category.CategoryUpdateReqDto;
-import keapoint.onlog.post.dto.category.CategoryDeleteReqDto;
+import keapoint.onlog.post.dto.category.PutCategoryUpdateReqDto;
+import keapoint.onlog.post.dto.category.DeleteCategoryReqDto;
 
 import java.util.UUID;
 
@@ -30,12 +30,11 @@ public class CategoryService {
     private final CategoryRepository categoryRepository;
 
     /**
-     * 새 카테고리를 생성합니다.
+     * 새 카테고리를 생성하는 메소드
      *
      * @param blogId 카테고리를 생성할 블로그의 ID
-     * @param dto 카테고리 생성 요청 데이터
+     * @param dto    카테고리 생성 요청 데이터
      * @return 생성된 카테고리 정보
-     * @throws BaseException 블로그 또는 토픽을 찾을 수 없거나, 이미 같은 이름의 카테고리가 존재하는 경우 발생
      */
     public CategoryDto createCategory(UUID blogId, PostCreateCategoryReqDto dto) throws BaseException {
         try {
@@ -77,26 +76,30 @@ public class CategoryService {
     }
 
     /**
-     * 기존 카테고리를 수정합니다.
+     * 기존 카테고리를 수정하는 메소드
      *
-     * @param dto 카테고리 수정 요청 데이터
+     * @param blogId 사용자의 블로그 식별자
+     * @param dto    카테고리 수정 요청 데이터
      * @return 수정된 카테고리 정보
-     * @throws BaseException 카테고리를 찾을 수 없거나, 이미 같은 이름의 카테고리가 존재하는 경우 발생
      */
     // 카테고리 수정 API
-    public CategoryDto updateCategory(CategoryUpdateReqDto dto) throws BaseException {
+    public CategoryDto updateCategory(UUID blogId, PutCategoryUpdateReqDto dto) throws BaseException {
         try {
+            // 사용자 조회
+            Blog blog = blogRepository.findById(blogId)
+                    .orElseThrow(() -> new BaseException(BaseErrorCode.BLOG_NOT_FOUND_EXCEPTION));
+
             // 카테고리 조회
             Category category = categoryRepository.findById(dto.getId())
                     .orElseThrow(() -> new BaseException(BaseErrorCode.CATEGORY_NOT_FOUND_EXCEPTION));
 
-            // 사용자가 해당 이름으로 카테고리를 가지고 있는지 조회
-            if (categoryRepository.findByNameAndCategoryOwner(dto.getName(), category.getCategoryOwner()).isPresent())
-                throw new BaseException(BaseErrorCode.ALREADY_CATEGORY_EXISTS_EXCEPTION);
+            // 사용자가 해당 카테고리를 가지고 있는지 조회
+            // 사용자가 만든 카테고리가 아니라면, UNAUTHORIZED_CATEGORY_ACCESS_EXCEPTION
+            if (blog.getCategories().contains(category))
+                throw new BaseException(BaseErrorCode.UNAUTHORIZED_CATEGORY_ACCESS_EXCEPTION);
 
             // 카테고리 이름 수정
             category.updateCategory(dto.getName());
-
             log.info("수정된 카테고리: " + category);
 
             // 수정된 카테고리 반환
@@ -113,22 +116,33 @@ public class CategoryService {
     }
 
     /**
-     * 기존 카테고리를 삭제합니다.
+     * 기존 카테고리를 삭제하는 메소드
      *
-     * @param dto 카테고리 삭제 요청 데이터
+     * @param blogId 사용자의 블로그 식별자
+     * @param dto    카테고리 삭제 요청 데이터
      * @return 삭제된 카테고리 정보
-     * @throws BaseException 카테고리를 찾을 수 없는 경우 발생
      */
-    public CategoryDto deleteCategory(CategoryDeleteReqDto dto) throws BaseException {
-        Category category = null;
+    public CategoryDto deleteCategory(UUID blogId, DeleteCategoryReqDto dto) throws BaseException {
         try {
+            // 사용자 조회
+            Blog blog = blogRepository.findById(blogId)
+                    .orElseThrow(() -> new BaseException(BaseErrorCode.BLOG_NOT_FOUND_EXCEPTION));
+
             // 카테고리 조회
-            category = categoryRepository.findById(dto.getId())
+            Category category = categoryRepository.findById(dto.getId())
                     .orElseThrow(() -> new BaseException(BaseErrorCode.CATEGORY_NOT_FOUND_EXCEPTION));
+
+            // 사용자가 해당 카테고리를 가지고 있는지 조회
+            // 사용자가 만든 카테고리가 아니라면, UNAUTHORIZED_CATEGORY_ACCESS_EXCEPTION
+            if (blog.getCategories().contains(category))
+                throw new BaseException(BaseErrorCode.UNAUTHORIZED_CATEGORY_ACCESS_EXCEPTION);
 
             // 카테고리 삭제
             categoryRepository.delete(category);
             log.info("삭제된 카테고리: " + category);
+
+            // 삭제된 카테고리 반환
+            return new CategoryDto(category);
 
         } catch (BaseException e) {
             log.error(e.getErrorCode().getMessage());
@@ -138,8 +152,5 @@ public class CategoryService {
             log.error(e.getMessage());
             throw new BaseException(BaseErrorCode.INTERNAL_SERVER_ERROR);
         }
-
-        // 삭제된 카테고리 반환
-        return new CategoryDto(category);
     }
 }
